--- server/src/main/java/com/unosquare/carmigo/dto/GrabUserAccessStatusDTO.java
@@ -1,11 +1,0 @@
-package com.unosquare.carmigo.dto;
-
-import lombok.Data;
-
-@Data
-public class GrabUserAccessStatusDTO {
-
-  private int id;
-
-  private String status;
-}
