+++ server/src/main/java/com/unosquare/carmigo/dto/request/JourneyRequest.java
@@ -1,4 +1,4 @@
-package com.unosquare.carmigo.model.request;
+package com.unosquare.carmigo.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -6,7 +6,7 @@ import java.time.Instant;
 import lombok.Data;
 
 @Data
-public class SearchJourneysRequest {
+public class JourneyRequest {
 
   @Positive
   private int locationIdFrom;
@@ -14,9 +14,9 @@ public class SearchJourneysRequest {
   @Positive
   private int locationIdTo;
 
-  @NotNull
-  private Instant dateTimeFrom;
+  @Positive
+  private int maxPassengers;
 
   @NotNull
-  private Instant dateTimeTo;
+  private Instant dateTime;
 }
