+++ server/src/main/java/com/unosquare/carmigo/dto/response/DistanceResponse.java
@@ -1,4 +1,4 @@
-package com.unosquare.carmigo.model.response;
+package com.unosquare.carmigo.dto.response;
 
 import lombok.Data;
