+++ server/src/main/java/com/unosquare/carmigo/dto/response/DriverResponse.java
@@ -1,6 +1,5 @@
-package com.unosquare.carmigo.model.response;
+package com.unosquare.carmigo.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
