+++ server/src/main/java/com/unosquare/carmigo/dto/response/UserAccessStatusResponse.java
@@ -1,4 +1,4 @@
-package com.unosquare.carmigo.model.response;
+package com.unosquare.carmigo.dto.response;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
