+++ server/src/main/java/com/unosquare/carmigo/service/PlatformUserService.java
@@ -11,8 +11,8 @@ import com.unosquare.carmigo.dto.PlatformUserDto;
 import com.unosquare.carmigo.entity.PlatformUser;
 import com.unosquare.carmigo.entity.UserAccessStatus;
 import com.unosquare.carmigo.exception.PatchException;
-import com.unosquare.carmigo.model.request.PlatformUserRequest;
-import com.unosquare.carmigo.model.response.PlatformUserResponse;
+import com.unosquare.carmigo.dto.request.PlatformUserRequest;
+import com.unosquare.carmigo.dto.response.PlatformUserResponse;
 import com.unosquare.carmigo.repository.PlatformUserRepository;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityManager;
