+++ server/src/test/java/com/unosquare/carmigo/service/AuthenticationServiceTest.java
@@ -10,8 +10,8 @@ import static org.mockito.Mockito.when;
 import com.flextrade.jfixture.FixtureAnnotations;
 import com.flextrade.jfixture.JFixture;
 import com.flextrade.jfixture.annotations.Fixture;
-import com.unosquare.carmigo.model.request.AuthenticationRequest;
-import com.unosquare.carmigo.model.response.AuthenticationResponse;
+import com.unosquare.carmigo.dto.request.AuthenticationRequest;
+import com.unosquare.carmigo.dto.response.AuthenticationResponse;
 import com.unosquare.carmigo.security.UserSecurityService;
 import com.unosquare.carmigo.util.JwtTokenService;
 import java.util.ArrayList;
