+++ server/src/test/java/com/unosquare/carmigo/service/DriverServiceTest.java
@@ -15,8 +15,8 @@ import com.flextrade.jfixture.JFixture;
 import com.flextrade.jfixture.annotations.Fixture;
 import com.unosquare.carmigo.entity.Driver;
 import com.unosquare.carmigo.entity.PlatformUser;
-import com.unosquare.carmigo.model.request.DriverRequest;
-import com.unosquare.carmigo.model.response.DriverResponse;
+import com.unosquare.carmigo.dto.request.DriverRequest;
+import com.unosquare.carmigo.dto.response.DriverResponse;
 import com.unosquare.carmigo.repository.DriverRepository;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityManager;
