+++ server/src/test/java/com/unosquare/carmigo/service/JourneyServiceTest.java
@@ -24,10 +24,10 @@ import com.unosquare.carmigo.entity.Location;
 import com.unosquare.carmigo.entity.Passenger;
 import com.unosquare.carmigo.exception.ResourceNotFoundException;
 import com.unosquare.carmigo.exception.UnauthorizedException;
-import com.unosquare.carmigo.model.request.DistanceRequest;
-import com.unosquare.carmigo.model.request.JourneyRequest;
-import com.unosquare.carmigo.model.request.SearchJourneysRequest;
-import com.unosquare.carmigo.model.response.JourneyResponse;
+import com.unosquare.carmigo.dto.request.DistanceRequest;
+import com.unosquare.carmigo.dto.request.JourneyRequest;
+import com.unosquare.carmigo.dto.request.SearchJourneysRequest;
+import com.unosquare.carmigo.dto.response.JourneyResponse;
 import com.unosquare.carmigo.openfeign.DistanceApi;
 import com.unosquare.carmigo.openfeign.DistanceHolder;
 import com.unosquare.carmigo.repository.JourneyRepository;
