+++ server/src/test/java/com/unosquare/carmigo/service/PassengerServiceTest.java
@@ -15,7 +15,7 @@ import com.flextrade.jfixture.JFixture;
 import com.flextrade.jfixture.annotations.Fixture;
 import com.unosquare.carmigo.entity.Passenger;
 import com.unosquare.carmigo.entity.PlatformUser;
-import com.unosquare.carmigo.model.response.PassengerResponse;
+import com.unosquare.carmigo.dto.response.PassengerResponse;
 import com.unosquare.carmigo.repository.PassengerRepository;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityManager;
