+++ server/src/test/java/com/unosquare/carmigo/service/PlatformUserServiceTest.java
@@ -21,8 +21,8 @@ import com.github.fge.jsonpatch.JsonPatch;
 import com.unosquare.carmigo.dto.PlatformUserDto;
 import com.unosquare.carmigo.entity.PlatformUser;
 import com.unosquare.carmigo.entity.UserAccessStatus;
-import com.unosquare.carmigo.model.request.PlatformUserRequest;
-import com.unosquare.carmigo.model.response.PlatformUserResponse;
+import com.unosquare.carmigo.dto.request.PlatformUserRequest;
+import com.unosquare.carmigo.dto.response.PlatformUserResponse;
 import com.unosquare.carmigo.repository.PlatformUserRepository;
 import com.unosquare.carmigo.util.PatchUtility;
 import com.unosquare.carmigo.util.ResourceUtility;
